main: Create the menu and use the Button Class()
The first menu is when the game starts, and will be 2 buttons (start and
exit). If the user click start the game will starts, if the user click
exit the game will quit.

A second menu is when the player already start the game and press the
key P of the keyboard. This action will generate a another menu where
exist 2 buttons (continue and exit)